refactor: 패키지 구조 변경 및 AuthExceptionHandler 삭제
- AuthExceptionHandler 삭제로 불필요한 코드 제거
- `reservationTime` 및 기타 관련 패키지의 naming을 `reservation.time`으로 수정
- 서비스 계층 패키지를 `application`으로 이동하여 구조 개선
