+++ src/main/java/roomescape/admin/application/AdminQueryService.java
@@ -1,0 +1,30 @@
+package roomescape.admin.application;
+
+import org.springframework.stereotype.Service;
+import roomescape.admin.domain.Admin;
+import roomescape.admin.domain.AdminRepository;
+import roomescape.common.exception.BusinessException;
+
+@Service
+public class AdminQueryService {
+
+    private final AdminRepository adminRepository;
+
+    public AdminQueryService(AdminRepository adminRepository) {
+        this.adminRepository = adminRepository;
+    }
+
+    public boolean isExistsByEmail(final String email) {
+        return adminRepository.existsByEmail(email);
+    }
+
+    public Admin findByEmail(final String email) {
+        return adminRepository.findByEmail(email)
+            .orElseThrow(() -> new BusinessException("관리자를 찾을 수 없습니다."));
+    }
+
+    public Admin findById(final Long id) {
+        return adminRepository.findById(id)
+            .orElseThrow(() -> new BusinessException("관리자를 찾을 수 없습니다."));
+    }
+}
