+++ src/main/java/roomescape/reservation/time/infrastructure/JpaReservationTimeRepository.java
@@ -1,0 +1,10 @@
+package roomescape.reservation.time.infrastructure;
+
+import java.util.List;
+import org.springframework.data.repository.CrudRepository;
+import roomescape.reservation.time.domain.ReservationTime;
+
+public interface JpaReservationTimeRepository extends CrudRepository<ReservationTime, Long> {
+
+    List<ReservationTime> findAll();
+}
