+++ src/main/java/roomescape/reservation/time/infrastructure/JpaReservationTimeRepositoryAdaptor.java
@@ -1,0 +1,42 @@
+package roomescape.reservation.time.infrastructure;
+
+import java.util.List;
+import java.util.Optional;
+import org.springframework.stereotype.Repository;
+import roomescape.reservation.time.domain.ReservationTime;
+import roomescape.reservation.time.domain.ReservationTimeRepository;
+
+@Repository
+public class JpaReservationTimeRepositoryAdaptor implements ReservationTimeRepository {
+
+    private final JpaReservationTimeRepository jpaReservationTimeRepository;
+
+    public JpaReservationTimeRepositoryAdaptor(JpaReservationTimeRepository jpaReservationTimeRepository) {
+        this.jpaReservationTimeRepository = jpaReservationTimeRepository;
+    }
+
+    @Override
+    public ReservationTime save(ReservationTime reservationTime) {
+        return jpaReservationTimeRepository.save(reservationTime);
+    }
+
+    @Override
+    public Optional<ReservationTime> findById(Long id) {
+        return jpaReservationTimeRepository.findById(id);
+    }
+
+    @Override
+    public boolean existsById(Long id) {
+        return jpaReservationTimeRepository.existsById(id);
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        jpaReservationTimeRepository.deleteById(id);
+    }
+
+    @Override
+    public List<ReservationTime> findAll() {
+        return jpaReservationTimeRepository.findAll();
+    }
+}
