+++ src/main/java/roomescape/reservation/time/presentation/dto/ReservationTimeRequest.java
@@ -1,0 +1,11 @@
+package roomescape.reservation.time.presentation.dto;
+
+import java.time.LocalTime;
+
+public record ReservationTimeRequest(LocalTime startAt) {
+    public ReservationTimeRequest {
+        if (startAt == null) {
+            throw new IllegalArgumentException("time은 null 일 수 없습니다.");
+        }
+    }
+}
