+++ src/main/java/roomescape/reservation/time/presentation/dto/ReservationTimeResponse.java
@@ -1,0 +1,10 @@
+package roomescape.reservation.time.presentation.dto;
+
+import java.time.LocalTime;
+import roomescape.reservation.time.domain.ReservationTime;
+
+public record ReservationTimeResponse(Long id, LocalTime startAt) {
+    public static ReservationTimeResponse from(final ReservationTime findReservationTime) {
+        return new ReservationTimeResponse(findReservationTime.getId(), findReservationTime.getStartAt());
+    }
+}
