+++ src/main/java/roomescape/reservation/time/presentation/dto/TimeConditionResponse.java
@@ -1,0 +1,6 @@
+package roomescape.reservation.time.presentation.dto;
+
+import java.time.LocalTime;
+
+public record TimeConditionResponse(Long id, LocalTime startAt, boolean alreadyBooked) {
+}
