--- src/main/java/roomescape/reservationTime/domain/ReservationTimeRepository.java
@@ -1,17 +1,0 @@
-package roomescape.reservationTime.domain;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ReservationTimeRepository {
-
-    ReservationTime save(ReservationTime reservationTime);
-
-    Optional<ReservationTime> findById(Long id);
-
-    void deleteById(Long id);
-
-    List<ReservationTime> findAll();
-
-    boolean existsById(Long id);
-}
