--- src/main/java/roomescape/reservationTime/presentation/dto/ReservationTimeResponse.java
@@ -1,10 +1,0 @@
-package roomescape.reservationTime.presentation.dto;
-
-import java.time.LocalTime;
-import roomescape.reservationTime.domain.ReservationTime;
-
-public record ReservationTimeResponse(Long id, LocalTime startAt) {
-    public static ReservationTimeResponse from(final ReservationTime findReservationTime) {
-        return new ReservationTimeResponse(findReservationTime.getId(), findReservationTime.getStartAt());
-    }
-}
