--- src/main/java/roomescape/reservationTime/presentation/dto/TimeConditionRequest.java
@@ -1,15 +1,0 @@
-package roomescape.reservationTime.presentation.dto;
-
-import java.time.LocalDate;
-
-public record TimeConditionRequest(LocalDate date, Long themeId) {
-    public TimeConditionRequest {
-        if (date == null) {
-            throw new IllegalArgumentException("date는 null 일 수 없습니다.");
-        }
-
-        if (themeId == null) {
-            throw new IllegalArgumentException("theme id는 null 일 수 없습니다.");
-        }
-    }
-}
